refactor : NumberGenerator를 interface화 하여 NumberGeneratorImpl 구현
