+++ src/main/java/lotto/domain/game/NumberGenerator.java
@@ -1,25 +1,12 @@
 package lotto.domain.game;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-public class NumberGenerator {
+public interface NumberGenerator {
 
-    private static final int MIN_NUMBER = 1;
-    private static final int MAX_NUMBER = 45;
+    int MIN_NUMBER = 1;
+    int MAX_NUMBER = 45;
 
-    private List<Integer> generateNumbers() {
-        return IntStream.rangeClosed(MIN_NUMBER, MAX_NUMBER)
-            .boxed()
-            .collect(Collectors.toList());
-    }
-
-    public List<Integer> shuffleNumbers() {
-        List<Integer> numbers = generateNumbers();
-        Collections.shuffle(numbers);
-        return numbers;
-    }
+    List<Integer> shuffleNumbers();
 
 }
