+++ src/test/java/lotto/LottoStatisticsTest.java
@@ -27,12 +27,7 @@ public class LottoStatisticsTest {
             "7,8,9,10,11,12"
         );
         LottoCount lottoCount = new LottoCount(payment, lottoManualCount);
-        NumberGenerator numberGenerator = new NumberGenerator(){
-            @Override
-            public List<Integer> shuffleNumbers() {
-                return List.of(1,2,3,4,5,6);
-            }
-        };
+        NumberGenerator numberGenerator = () -> List.of(1,2,3,4,5,6);
         Lottos lottos = new Lottos(lottoCount, lottosManual, numberGenerator);
 
         String lottoValue = "1,2,3,4,5,6";
@@ -54,12 +49,8 @@ public class LottoStatisticsTest {
             "7,8,9,10,11,12"
         );
         LottoCount lottoCount = new LottoCount(payment, lottoManualCount);
-        NumberGenerator numberGenerator = new NumberGenerator(){
-            @Override
-            public List<Integer> shuffleNumbers() {
-                return List.of(40,41,42,43,44,45);
-            }
-        };
+        NumberGenerator numberGenerator = () -> List.of(40,41,42,43,44,45);
+
         Lottos lottos = new Lottos(lottoCount, lottosManual, numberGenerator);
 
         String lottoValue = "1,2,3,4,5,8";
